+++ Step 1. Setup HyperV for Lab.ps1
@@ -6,12 +6,14 @@ If($PSScriptRoot.ToString().length -eq 0)
      Break
 }
 Else{
-    Write-Host ("Loading configuration file first...") -ForegroundColor Yellow
+    Write-Host ("Loading configuration file first...") -ForegroundColor Yellow -NoNewline
     . "$PSScriptRoot\configs.ps1" -NoAzureCheck
+    Write-Host "Done" -ForegroundColor Green
 }
 #endregion
 
 #install hyper-V feature
+Write-Host ("Enabling Hyper-V role...") -NoNewline
 If( (Get-WindowsOptionalFeature -Online -FeatureName Microsoft-Hyper-V).State -ne 'Enabled'){
     $HyperVResult = Enable-WindowsOptionalFeature -Online -FeatureName Microsoft-Hyper-V -All
     If($HyperVResult.RestartNeeded -eq $true){
@@ -22,11 +24,10 @@ Else{
     Write-Host ("Hyper-V is already installed.") -ForegroundColor Green
 }
 
-
+#region Configure Hyper-V settings
+Write-Host ("Setting up Hyper-V...") -NoNewline
 #Grab physical network
 $FastestPhysicalAdapter = (Get-NetAdapter -Physical | Where-Object {$_.Status -eq 'Up' -and $_.InterfaceDescription -notmatch "Xbox"} | Sort-Object $_.LinkSpeed | Select-Object -First 1).Name
-
-#region Configure Hyper-V settings
 If( ($HyperVConfig.ChangeLocation) -and ((Get-VMHost).VirtualMachinePath -ne $HyperVConfig.VirtualMachineLocation) -or ((Get-VMHost).VirtualHardDiskPath -ne $HyperVConfig.VirtualHardDiskLocation) ){
     New-Item $HyperVConfig.VirtualMachineLocation -ItemType Directory -ErrorAction SilentlyContinue
     New-Item $HyperVConfig.VirtualHardDiskLocation -ItemType Directory -ErrorAction SilentlyContinue
@@ -35,28 +36,35 @@ If( ($HyperVConfig.ChangeLocation) -and ((Get-VMHost).VirtualMachinePath -ne $Hy
 Else{
     Set-VMHost -EnableEnhancedSessionMode:$HyperVConfig.EnableSessionMode
 }
+Write-Host "Done" -ForegroundColor Green
 #endregion
 
-#Private Switches
-$HyperVSwitches = Get-VMSwitch
-
 #region Create External Switch
+Write-Host ("Configuring Hyper-V External switch...") -NoNewline
+$HyperVSwitches = Get-VMSwitch
 If ($null -eq ($HyperVSwitches | Where SwitchType -eq 'External') ) {
     New-VMSwitch -Name 'External' -NetAdapterName $FastestPhysicalAdapter -AllowManagementOS $true -Notes 'External Switch'
 }
+
+Write-Host "Done" -ForegroundColor Green
 #endregion
 
 $i = 1
 #TEST $Subnet = $VyOSConfig.LocalSubnetPrefix.GetEnumerator() | Sort Name |Select -first 1
-
-Foreach($Subnet in $VyOSConfig.LocalSubnetPrefix.GetEnumerator() | Sort Name){
+Foreach($Subnet in $VyOSConfig.LocalSubnetPrefix.GetEnumerator() | Sort Name)
+{
     $NetworkName = ($vYosConfig.NetPrefix +' ' + $i + ' - ' + $Subnet.Name)
+    Write-Host ("Configuring Hyper-V internal switch: [{0}]..." -f $NetworkName) -NoNewline
     $Description = ("{2} for {1}: {0}" -f $Subnet.Name,$VyOSConfig.LocalSubnetPrefix[$Subnet.Name],$vYosConfig.NetPrefix)
-    If($NetworkName -notin ($HyperVSwitches | Where SwitchType -eq 'Private') ){
-        New-VMSwitch -Name $NetworkName -SwitchType Private -Notes $Description
+    If( $HyperVSwitches | Where Name -eq $NetworkName ){
+        Write-Host ("Network already exists. Skipping creation.") -ForegroundColor Green
     }
     Else{
-        Write-Host ("{0} Network already exists. Skipping creation." -f $NetworkName) -ForegroundColor Yellow
+        New-VMSwitch -Name $NetworkName -SwitchType Private -Notes $Description
+        Write-Host "Done" -ForegroundColor Green
     }
     $i++
 }
+
+Write-Host ("Done configuring Hyper-V") -ForegroundColor Green
+Write-Host "--------------------------------------------------" -ForegroundColor Green